Migrate timestamps to int64 and fix tests
This commit fixes all the tests that were not passing

Also migrates to latest version of carbonapi_v3_pb

+++ date/date_test.go
@@ -41,7 +41,7 @@ func TestDateParamToEpoch(t *testing.T) {
 			panic(fmt.Sprintf("error parsing time: %q: %v", tt.output, err))
 		}
 
-		want := uint32(ts.Unix())
+		want := int64(ts.Unix())
 		if got != want {
 			t.Errorf("dateParamToEpoch(%q, 0)=%v, want %v", tt.input, got, want)
 		}
