+++ expr/functions/averageSeries/function.go
@@ -25,7 +25,7 @@ func New(configFile string) []interfaces.FunctionMetadata {
 }
 
 // averageSeries(*seriesLists)
-func (f *averageSeries) Do(e parser.Expr, from, until uint32, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
+func (f *averageSeries) Do(e parser.Expr, from, until int64, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
 	args, err := helper.GetSeriesArgsAndRemoveNonExisting(e, from, until, values)
 	if err != nil {
 		return nil, err
