+++ expr/functions/averageSeriesWithWildcards/function.go
@@ -29,7 +29,7 @@ func New(configFile string) []interfaces.FunctionMetadata {
 }
 
 // averageSeriesWithWildcards(seriesLIst, *position)
-func (f *averageSeriesWithWildcards) Do(e parser.Expr, from, until uint32, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
+func (f *averageSeriesWithWildcards) Do(e parser.Expr, from, until int64, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
 	/* TODO(dgryski): make sure the arrays are all the same 'size'
 	   (duplicated from sumSeriesWithWildcards because of similar logic but aggregation) */
 	args, err := helper.GetSeriesArg(e.Args()[0], from, until, values)
