+++ expr/functions/cactiStyle/function.go
@@ -29,7 +29,7 @@ func New(configFile string) []interfaces.FunctionMetadata {
 }
 
 // cactiStyle(seriesList, system=None, units=None)
-func (f *cactiStyle) Do(e parser.Expr, from, until uint32, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
+func (f *cactiStyle) Do(e parser.Expr, from, until int64, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
 	// Get the series data
 	original, err := helper.GetSeriesArg(e.Args()[0], from, until, values)
 	if err != nil {
