+++ expr/functions/consolidateBy/function.go
@@ -26,7 +26,7 @@ func New(configFile string) []interfaces.FunctionMetadata {
 }
 
 // consolidateBy(seriesList, aggregationMethod)
-func (f *consolidateBy) Do(e parser.Expr, from, until uint32, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
+func (f *consolidateBy) Do(e parser.Expr, from, until int64, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
 	arg, err := helper.GetSeriesArg(e.Args()[0], from, until, values)
 	if err != nil {
 		return nil, err
