+++ expr/functions/holtWintersConfidenceBands/function.go
@@ -28,7 +28,7 @@ func New(configFile string) []interfaces.FunctionMetadata {
 	return res
 }
 
-func (f *holtWintersConfidenceBands) Do(e parser.Expr, from, until uint32, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
+func (f *holtWintersConfidenceBands) Do(e parser.Expr, from, until int64, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
 	var results []*types.MetricData
 	args, err := helper.GetSeriesArg(e.Args()[0], from-7*86400, until, values)
 	if err != nil {
