+++ expr/functions/moving/function.go
@@ -29,7 +29,7 @@ func New(configFile string) []interfaces.FunctionMetadata {
 }
 
 // movingXyz(seriesList, windowSize)
-func (f *moving) Do(e parser.Expr, from, until uint32, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
+func (f *moving) Do(e parser.Expr, from, until int64, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
 	var n int
 	var err error
 
@@ -58,7 +58,7 @@ func (f *moving) Do(e parser.Expr, from, until uint32, values map[parser.MetricR
 
 	start := from
 	if scaleByStep {
-		start -= uint32(n)
+		start -= int64(n)
 	}
 
 	arg, err := helper.GetSeriesArg(e.Args()[0], start, until, values)
