+++ expr/functions/multiplySeriesWithWildcards/function.go
@@ -30,7 +30,7 @@ func New(configFile string) []interfaces.FunctionMetadata {
 }
 
 // multiplySeriesWithWildcards(seriesList, *position)
-func (f *multiplySeriesWithWildcards) Do(e parser.Expr, from, until uint32, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
+func (f *multiplySeriesWithWildcards) Do(e parser.Expr, from, until int64, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
 	/* TODO(dgryski): make sure the arrays are all the same 'size'
 	   (duplicated from sumSeriesWithWildcards because of similar logic but multiplication) */
 	args, err := helper.GetSeriesArg(e.Args()[0], from, until, values)
